reorganize buildout
Split buildout.cfg into base.cfg and development.cfg. The former holds the main buildout, the latter development specific settings. As of now only some deveopment addons are configured for installation. The absence of an buildout.cfg file means that it is now impossible to run buildout with the wrong configuration file on the production server.

Also clean up the buildout in general. Removes some redundant settings and the productdistros part. We don't use old-style products anymore.