sbo.theme: use own css grid
The Plone one uses percentages to specify widths and positions of the cells in the grid. This leads to theoretically fractional sizes which are interpreded differently by the browsers. Our grind uses precalculated pixels sizes to work arround the issue.

Needed for the new gallery views. They depend one exact width of the main content column or the layout will break.